Adding network performance tooling
UWPiPerf is a tool that provides server and client sides to a network
performance test. The client appilcation provides test configurations
and can be further configured from within the code.
First step should be to define the IoT Hub connection settings on both
the Server and the Client.
GPS location data is not gathered with this build.